@@ -13,10 +13,11 @@
   },
   "dependencies": {
     "chart.js": "3.7.0",
+    "idb": "^7.0.1",
     "less": "^4.1.2",
-    "react": "17.0.2",
+    "react": "18.1.0",
     "react-chartjs-2": "^4.0.1",
-    "react-dom": "17.0.2",
+    "react-dom": "18.1.0",
     "react-router-dom": "6.2.1",
     "semantic-ui-react": "^2.0.4",
     "twemoji": "13.1.0"
