@@ -2,9 +2,9 @@ import { useParams } from 'react-router-dom';
 
 // Components
 import { School as SchoolInfo } from '@/components/School';
-import { NotFound } from '@/pages/404';
+import NotFound from '@/pages/404';
 
-export const School: React.FC = () => {
+const School: React.FC = () => {
   const params = useParams();
   const id = Number(params.id) ?? -1;
   const school = OIerDb.schools.find((school) => school.id === id);
@@ -18,3 +18,5 @@ export const School: React.FC = () => {
     </>
   );
 };
+
+export default School;
